% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_frequency.R
\name{ap_frequency}
\alias{ap_frequency}
\title{ap_frequency}
\usage{
ap_frequency(
  x,
  step,
  ap_threshold = -1,
  iStep_window = 6564:11561,
  baseline_window = 1:6000
)
}
\arguments{
\item{x}{data frame containing a time column, voltage trace column(s), and current trace column(s).}

\item{step}{user defined current step (pA) that ap frequency will be calculated}

\item{ap_threshold}{user defined voltage threshold for action potentials. default it -20mV (Derek's default). adjust acorrdingly.}

\item{iStep_window}{indices representing the entire current step. defaults to ~650ms - 1250ms window, which is Derek's default. adjust your window accordingly.}

\item{baseline_window}{indices representing the baseline window that is used for calculation. defaults to first 600ms of each sweep, which is Derek's default. adjust your window accordingly.}
}
\value{
returns action potential frequency in Hz
}
\description{
This function returns the action potential frequency from a user defined current step from a current step experiment.
}
\author{
Derek Merryweather
}
