% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineSub.R
\name{baselineSub}
\alias{baselineSub}
\title{baselineSub}
\usage{
baselineSub(x, baseline_window)
}
\arguments{
\item{x}{Either a data frame or vector of values to be normalized}

\item{baseline_window}{Vector of the baseline window}
}
\value{
returns the baseline subtracted x
}
\description{
This function is used to subtract values from a defined baseline window
}
\author{
Derek Merryweather
}
