% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{getmode}
\usage{
getmode(x)
}
\arguments{
\item{x}{must be a numeric vector}
}
\value{
returns a single value that is the mode of vector x
}
\description{
This function gets the mode of a vector x
}
\author{
Derek Merryweather
}
