% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{input_output}
\alias{input_output}
\title{input_output}
\usage{
input_output(
  x,
  ap_threshold = -20,
  iStep_window = 6564:11561,
  baseline_window = 1:6000
)
}
\arguments{
\item{x}{data frame containing a time column, voltage trace column(s), and current trace column(s).}

\item{ap_threshold}{user defined voltage threshold for action potentials. default it -20mV (Derek's default). adjust acorrdingly.}

\item{iStep_window}{indices representing the entire current step. defaults to ~650ms - 1250ms window, which is Derek's default. adjust your window accordingly.}

\item{baseline_window}{indices representing the baseline window that is used for calculation. defaults to first 600ms of each sweep, which is Derek's default. adjust your window accordingly.}
}
\value{
returns a data frame with two columns: input (pA) and output (APnum).
}
\description{
This function is used to look at the input / output relationship of a current step experiment.
}
\author{
Derek Merryweather
}
