% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interSpikeInterval.R
\name{interSpikeInterval}
\alias{interSpikeInterval}
\title{interSpikeInterval}
\usage{
interSpikeInterval(
  x,
  step,
  ap_threshold = -20,
  iStep_window = 6564:11561,
  baseline_window = 1:6000
)
}
\arguments{
\item{x}{data frame containing a time column, voltage trace column(s), and current trace column(s).}

\item{step}{user defined current step to calculate ISI}

\item{ap_threshold}{threshold voltage for detecting action potential. default is -20mV.}

\item{iStep_window}{indices representing the entire current step. defaults to ~650ms - 1250ms window, which is Derek's default. adjust your window accordingly.}

\item{baseline_window}{indices representing the baseline window that is used for calculation. defaults to first 600ms of each sweep, which is Derek's default. adjust your window accordingly.}
}
\value{
returns vector of ISI values in ms for user defined current step. first element will be the spike delay time
}
\description{
finds the inter spike interval for user defined current step
}
\author{
Derek Merryweather
}
