% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rheobase.R
\name{rheobase}
\alias{rheobase}
\title{rheobase}
\usage{
rheobase(
  x,
  iStep_window = 6564:11561,
  baseline_window = 1:6000,
  ap_threshold = -20
)
}
\arguments{
\item{x}{dataframe containing time column, voltage trace column(s), and current trace column(s)}

\item{iStep_window}{indices representing the entire current step. defaults to ~650ms - 1250ms window, which is Derek's default. adjust your window accordingly.}

\item{baseline_window}{indices representing the baseline window that is used for calculation. defaults to first 600ms of each sweep, which is Derek's default. adjust your window accordinlgy.}

\item{ap_threshold}{threshold voltage for detecting action potential. default is -20mV.}
}
\value{
returns a single numeric value that is the rheobase of current step experiment
}
\description{
function to extract rheobase value from current step experiment
}
\author{
Derek Merryweather
}
