% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sag.R
\name{sag}
\alias{sag}
\title{sag}
\usage{
sag(
  x,
  step,
  iStep_window = 6564:11561,
  baseline_window = 1:6000,
  vInfinity_window = 10001:11561,
  sag_window = 6564:8001
)
}
\arguments{
\item{x}{data frame containing a time column, voltage trace column(s), and current trace column(s).}

\item{step}{user defined current step (pA) that sag ratio will be calculated}

\item{iStep_window}{indices representing the entire current step. defaults to ~650ms - 1250ms window, which is Derek's default. adjust your window accordingly.}

\item{baseline_window}{indices representing the baseline window that is used for calculation. defaults to first 600ms, which is Derek's default. adjust your window accordingly.}

\item{vInfinity_window}{indices representing Vinfinity window that is used for calculation. defaults to last 100ms of iStep_window, which is Derek's default. adjust your window accordingly.}

\item{sag_window}{indices representing the sag window that is used for calculation. defaults to ~first 100ms of iStep_window, which is Derek's default. adjust your window accordingly.}
}
\value{
returns sag ratio value
}
\description{
This function is used to find the sag ratio of a user defined current step from a current step experiment
}
\author{
Derek Merryweather
}
