% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_type.R
\name{spike_type}
\alias{spike_type}
\title{spike_type}
\usage{
spike_type(
  x,
  ap_threshold = -1,
  iStep_window = 6564:11561,
  baseline_window = 1:6000,
  vInfinity_window = 10001:11561,
  burst_width = 150
)
}
\arguments{
\item{x}{data frame containing a time column, voltage trace column(s), and current trace column(s).}

\item{ap_threshold}{user defined voltage threshold for action potentials. default it -20mV (Derek's default). adjust acorrdingly.}

\item{iStep_window}{indices representing the entire current step. defaults to ~650ms - 1250ms window, which is Derek's default. adjust your window accordingly.}

\item{baseline_window}{indices representing the baseline window that is used for calculation. defaults to first 600ms, which is Derek's default. adjust your window accordingly.}

\item{vInfinity_window}{indices representing Vinfinity window that is used for calculation. defaults to last 100ms of iStep_window, which is Derek's default. adjust your window accordingly.}

\item{burst_width}{user defined threshold value indicating how far apart in time each ap is in a "burst" of aps. default set to 150 (150=15ms), which is Derek's default. adjust your window accordingly.}
}
\value{
returns character "regular" or "burst"
}
\description{
This function is used to find the spike type of a neuron (burst or regular) from the rheobase trace of a current step experiment.
}
\author{
Derek Merryweather
}
