% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_width.R
\name{spike_width}
\alias{spike_width}
\title{spike_width}
\usage{
spike_width(
  x,
  ap_threshold = -30,
  iStep_window = 6564:11561,
  baseline_window = 1:6000
)
}
\arguments{
\item{x}{data frame containing a time column, voltage trace column(s), and current trace column(s).}

\item{ap_threshold}{threshold voltage for detecting action potential.}

\item{iStep_window}{indices representing the entire current step. defaults to ~650ms - 1250ms window, which is Derek's default. adjust your window accordingly.}

\item{baseline_window}{indices representing the baseline window that is used for calculation. defaults to first 600ms of each sweep, which is Derek's default. adjust your window accordingly.}
}
\value{
returns the mean action potential width.
}
\description{
function that finds with average width of action potentials during a current step trace
}
\author{
Derek Merryweather
}
